requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
python-dotenv>=1.0.0
docker>=6.0.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
gunicorn>=20.1.0
psycopg2>=2.9.0
